py attic: Move attic unittests
