+++ bindings/pydrake/attic/multibody/BUILD.bazel
@@ -11,6 +11,7 @@ load(
 load(
     "@drake//tools/skylark:drake_py.bzl",
     "drake_py_library",
+    "drake_py_unittest",
 )
 
 package(default_visibility = [
@@ -26,7 +27,7 @@ drake_py_library(
     srcs = ["__init__.py"],
     imports = PACKAGE_INFO.py_imports,
     deps = [
-        "//bindings/pydrake:module_py",
+        "//bindings/pydrake/attic:module_py",
     ],
 )
 
@@ -159,8 +160,52 @@ install(
     deps = get_drake_py_installs(PY_LIBRARIES_WITH_INSTALL),
 )
 
-# NOTE: Unittests on `attic` should not move until the forwarding modules
-# outside of the attic are deprecated!
+drake_py_unittest(
+    name = "rigid_body_plant_test",
+    data = ["//examples/pendulum:models"],
+    deps = [
+        ":rigid_body_plant_py",
+        ":rigid_body_tree_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "shapes_test",
+    data = ["//examples/quadrotor:models"],
+    deps = [
+        ":shapes_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "rigid_body_tree_test",
+    data = [
+        "//examples/atlas:models",
+        "//examples/pendulum:models",
+        "//examples/simple_four_bar:models",
+    ],
+    deps = [
+        ":collision_py",
+        ":joints_py",
+        ":parsers_py",
+        ":rigid_body_py",
+        ":rigid_body_tree_py",
+        ":shapes_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "parsers_test",
+    data = [
+        "//examples/acrobot:models",
+        "//examples/atlas:models",
+        "//examples/pr2:models",
+    ],
+    deps = [
+        ":parsers_py",
+        ":rigid_body_tree_py",
+    ],
+)
 
 add_lint_tests(
     cpplint_data = ["//bindings/pydrake:.clang-format"],
