+++ bindings/pydrake/attic/multibody/test/rigid_body_plant_test.py
