+++ bindings/pydrake/attic/multibody/test/rigid_body_tree_test.py
