+++ bindings/pydrake/attic/multibody/test/shapes_test.py
