+++ bindings/pydrake/attic/solvers/BUILD.bazel
@@ -11,6 +11,7 @@ load(
 load(
     "@drake//tools/skylark:drake_py.bzl",
     "drake_py_library",
+    "drake_py_unittest",
 )
 
 package(default_visibility = [
@@ -26,7 +27,7 @@ drake_py_library(
     srcs = ["__init__.py"],
     imports = PACKAGE_INFO.py_imports,
     deps = [
-        "//bindings/pydrake:module_py",
+        "//bindings/pydrake/attic:module_py",
     ],
 )
 
@@ -69,8 +70,26 @@ install(
     deps = get_drake_py_installs(PY_LIBRARIES_WITH_INSTALL),
 )
 
-# NOTE: Unittests on `attic` should not move until the forwarding modules
-# outside of the attic are deprecated!
+drake_py_unittest(
+    name = "pr2_ik_test",
+    data = ["//examples/pr2:models"],
+    tags = ["snopt"],
+    deps = [
+        ":ik_py",
+        "//bindings/pydrake/attic/multibody:parsers_py",
+        "//bindings/pydrake/attic/multibody:rigid_body_tree_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "rbt_ik_test",
+    data = ["//examples/pendulum:models"],
+    tags = ["snopt"],
+    deps = [
+        ":ik_py",
+        "//bindings/pydrake/attic/multibody:rigid_body_tree_py",
+    ],
+)
 
 add_lint_tests(
     cpplint_data = ["//bindings/pydrake:.clang-format"],
