+++ bindings/pydrake/attic/solvers/test/pr2_ik_test.py
