+++ bindings/pydrake/attic/solvers/test/rbt_ik_test.py
