+++ bindings/pydrake/attic/systems/BUILD.bazel
@@ -11,6 +11,7 @@ load(
 load(
     "@drake//tools/skylark:drake_py.bzl",
     "drake_py_library",
+    "drake_py_unittest",
 )
 
 package(default_visibility = [
@@ -26,7 +27,7 @@ drake_py_library(
     srcs = ["__init__.py"],
     imports = PACKAGE_INFO.py_imports,
     deps = [
-        "//bindings/pydrake:module_py",
+        "//bindings/pydrake/attic:module_py",
     ],
 )
 
@@ -84,8 +85,27 @@ install(
     deps = get_drake_py_installs(PY_LIBRARIES_WITH_INSTALL),
 )
 
-# NOTE: Unittests on `attic` should not move until the forwarding modules
-# outside of the attic are deprecated!
+drake_py_unittest(
+    name = "controllers_test",
+    data = [
+        "//manipulation/models/iiwa_description:models",
+    ],
+    deps = [
+        ":controllers_py",
+        "//bindings/pydrake/multibody:rigid_body_tree_py",
+    ],
+)
+
+drake_py_unittest(
+    name = "sensors_test",
+    data = [
+        "//systems/sensors:test_models",
+    ],
+    deps = [
+        ":sensors_py",
+        "//bindings/pydrake/systems:sensors_py",
+    ],
+)
 
 add_lint_tests(
     cpplint_data = ["//bindings/pydrake:.clang-format"],
