+++ bindings/pydrake/solvers/BUILD.bazel
@@ -172,29 +172,6 @@ drake_py_unittest(
     ],
 )
 
-drake_py_test(
-    name = "pr2_ik_test",
-    # TODO(jwnimmer-tri) Port this to drake_py_unittest.
-    allow_import_unittest = True,
-    data = ["//examples/pr2:models"],
-    tags = ["snopt"],
-    deps = [
-        ":ik_py",
-        "//bindings/pydrake/multibody:parsers_py",
-        "//bindings/pydrake/multibody:rigid_body_tree_py",
-    ],
-)
-
-drake_py_unittest(
-    name = "rbt_ik_test",
-    data = ["//examples/pendulum:models"],
-    tags = ["snopt"],
-    deps = [
-        ":ik_py",
-        "//bindings/pydrake/multibody:rigid_body_tree_py",
-    ],
-)
-
 add_lint_tests(
     cpplint_data = ["//bindings/pydrake:.clang-format"],
     enable_clang_format_lint = True,
