+++ bindings/pydrake/systems/_controllers_extra.py
@@ -1,1 +1,5 @@
-from pydrake.attic.systems.controllers import *
+# Only try to import attic symbols if they're available.
+try:
+    from pydrake.attic.systems.controllers import *
+except ImportError:
+    pass
