+++ bindings/pydrake/systems/_sensors_extra.py
@@ -1,6 +1,5 @@
-# Due to dependency cycle, only try to import attic symbols if they're
-# available.
+# Only try to import attic symbols if they're available.
 try:
     from pydrake.attic.systems.sensors import *
-except ImportError as e:
+except ImportError:
     pass
